creating the club controller and first views

+++ database/migrations/2022_09_27_145428_create_member_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('email');
             $table->string('dni');
-            $table->integer('salary');
+            $table->float('salary');
             $table->unsignedBigInteger('type_id');
             $table->unsignedBigInteger('club_id')->nullable;
             $table->string('photo');
